+++ src/sessions/screen-recording-session-handler.ts
@@ -73,6 +73,21 @@ export default class ScreenRecordingSessionHandler extends BaseSessionHandler {
   }
 
   private updateScreenRecordingMetadatas (session: ScreenRecordingMediaSession, metadatas: ScreenRecordingMetadata[]) {
+    metadatas.forEach((meta) => {
+      // adding any here because I don't want to add this to the public interface
+      (meta as any)._trackId = meta.trackId;
+
+      const transceiver = session.pc.getTransceivers()
+        .find((transceiver) => transceiver.sender.track?.id === meta.trackId);
+
+      if (!transceiver) {
+        this.log('warn', 'Failed to find transceiver for screen recording track', { conversationId: session.conversationId, sessionId: session.id, trackId: meta.trackId });
+        return;
+      }
+
+      meta.trackId = transceiver.mid;
+    });
+    
     const data = JSON.stringify({
       participantJid: getBareJid(this.sdk),
       metaData: metadatas,
@@ -80,8 +95,15 @@ export default class ScreenRecordingSessionHandler extends BaseSessionHandler {
       conversationId: session.conversationId
     });
 
-    return requestApi.call(this.sdk, '/recordings/screensessions/metadata', {
+    let url = '/recordings/screensessions/metadata';
+    const { accessToken, jwt } = this.sdk._config;
+    if (this.sdk.isJwtAuth) {
+      url += '/backgroundassistant';
+    }
+
+    return requestApi.call(this.sdk, url, {
       method: 'post',
+      authToken: accessToken || jwt,
       data
     });
   }